modified Some Old apps

+++ Quera/Class/Exercises/HW3/HW3_Q2_402170516.cpp
@@ -17,7 +17,7 @@ int main()
 }
 void string_multiply(int n, string ch)
 {
-  for (size_t i = 0; i < n; i++)
+  for (int i = 0; i < n; i++)
   {
     cout << ch;
   }
